Scope reply events to current portal & quoted user
When bridging a Signal quote to a Matrix reply, consider events only
from the portal in which the reply is being bridged when searching for
the replied-to message.

Also refuse to bridge the quote if there is no puppet for the Signal
user who sent the message that was quoted.